+++ apm-sniffer/apm-sdk-plugin/jedis-plugins/jedis-2.x-3.x-plugin/src/main/java/org/apache/skywalking/apm/plugin/jedis/v3/JedisPluginConfig.java
@@ -19,6 +19,10 @@ package org.apache.skywalking.apm.plugin.jedis.v3;
 
 import org.apache.skywalking.apm.agent.core.boot.PluginConfig;
 
+import java.util.Arrays;
+import java.util.HashSet;
+import java.util.Set;
+
 public class JedisPluginConfig {
     public static class Plugin {
         @PluginConfig(root = JedisPluginConfig.class)
@@ -35,6 +39,119 @@ public class JedisPluginConfig {
              * Set a negative number to save specified length of parameter string to the tag.
              */
             public static int REDIS_PARAMETER_MAX_LENGTH = 128;
+
+            /**
+             * Operation represent a cache span is "write" or "read" action , and "op"(operation) is tagged with key "cache.op" usually
+             * This config term define which command should be converted to write Operation .
+             *
+             * @see org.apache.skywalking.apm.agent.core.context.tag.Tags#CACHE_OP
+             * @see JedisMethodInterceptor#parseOperation(String)
+             */
+            public static Set<String> OPERATION_MAPPING_WRITE = new HashSet<>(Arrays.asList(
+                    "getset",
+                    "set",
+                    "setbit",
+                    "setex ",
+                    "setnx ",
+                    "setrange",
+                    "strlen ",
+                    "mset",
+                    "msetnx ",
+                    "psetex",
+                    "incr ",
+                    "incrby ",
+                    "incrbyfloat",
+                    "decr ",
+                    "decrby ",
+                    "append ",
+                    "hmset",
+                    "hset",
+                    "hsetnx ",
+                    "hincrby",
+                    "hincrbyfloat",
+                    "hdel",
+                    "rpoplpush",
+                    "rpush",
+                    "rpushx",
+                    "lpush",
+                    "lpushx",
+                    "lrem",
+                    "ltrim",
+                    "lset",
+                    "brpoplpush",
+                    "linsert",
+                    "sadd",
+                    "sdiff",
+                    "sdiffstore",
+                    "sinterstore",
+                    "sismember",
+                    "srem",
+                    "sunion",
+                    "sunionstore",
+                    "sinter",
+                    "zadd",
+                    "zincrby",
+                    "zinterstore",
+                    "zrange",
+                    "zrangebylex",
+                    "zrangebyscore",
+                    "zrank",
+                    "zrem",
+                    "zremrangebylex",
+                    "zremrangebyrank",
+                    "zremrangebyscore",
+                    "zrevrange",
+                    "zrevrangebyscore",
+                    "zrevrank",
+                    "zunionstore",
+                    "xadd",
+                    "xdel",
+                    "del",
+                    "xtrim"
+            ));
+            /**
+             * Operation represent a cache span is "write" or "read" action , and "op"(operation) is tagged with key "cache.op" usually
+             * This config term define which command should be converted to write Operation .
+             *
+             * @see org.apache.skywalking.apm.agent.core.context.tag.Tags#CACHE_OP
+             * @see JedisMethodInterceptor#parseOperation(String)
+             */
+            public static Set<String> OPERATION_MAPPING_READ = new HashSet<>(Arrays.asList(
+                    "getrange",
+                    "getbit ",
+                    "mget",
+                    "hvals",
+                    "hkeys",
+                    "hlen",
+                    "hexists",
+                    "hget",
+                    "hgetall",
+                    "hmget",
+                    "blpop",
+                    "brpop",
+                    "lindex",
+                    "llen",
+                    "lpop",
+                    "lrange",
+                    "rpop",
+                    "scard",
+                    "srandmember",
+                    "spop",
+                    "sscan",
+                    "smove",
+                    "zlexcount",
+                    "zscore",
+                    "zscan",
+                    "zcard",
+                    "zcount",
+                    "xget",
+                    "get",
+                    "xread",
+                    "xlen",
+                    "xrange",
+                    "xrevrange"
+            ));
+
         }
     }
 }