+++ apm-sniffer/optional-plugins/ehcache-2.x-plugin/src/main/java/org/apache/skywalking/apm/plugin/ehcache/v2/EhcacheLockInterceptor.java
@@ -29,38 +29,40 @@ import org.apache.skywalking.apm.network.trace.component.ComponentsDefine;
 
 import java.lang.reflect.Method;
 
+import static org.apache.skywalking.apm.plugin.ehcache.v2.EncacheOperationConvertor.parseOperation;
 import static org.apache.skywalking.apm.plugin.ehcache.v2.define.EhcachePluginInstrumentation.LOCK_ENHANCE_METHOD_SUFFIX;
 
 public class EhcacheLockInterceptor implements InstanceMethodsAroundInterceptor {
 
     @Override
     public void beforeMethod(EnhancedInstance objInst, Method method, Object[] allArguments, Class<?>[] argumentsTypes,
-        MethodInterceptResult result) throws Throwable {
+                             MethodInterceptResult result) throws Throwable {
         EhcacheEnhanceInfo enhanceInfo = (EhcacheEnhanceInfo) objInst.getSkyWalkingDynamicField();
 
         String operateName = method.getName()
-                                   .substring(0, method.getName().length() - LOCK_ENHANCE_METHOD_SUFFIX.length());
+                .substring(0, method.getName().length() - LOCK_ENHANCE_METHOD_SUFFIX.length());
 
         AbstractSpan span = ContextManager.createLocalSpan("Ehcache/" + operateName + "/" + enhanceInfo.getCacheName());
         span.setComponent(ComponentsDefine.EHCACHE);
         SpanLayer.asCache(span);
-
-        Object element = allArguments[0];
-        if (element != null) {
-            Tags.DB_STATEMENT.set(span, element.toString());
+        if (allArguments != null && allArguments.length > 0 && allArguments[0] instanceof String) {
+            Tags.CACHE_KEY.set(span, enhanceInfo.getCacheName() + "." + allArguments[0]);
         }
+        Tags.CACHE_TYPE.set(span, "Ehcache");
+        Tags.CACHE_CMD.set(span, operateName);
+        parseOperation(operateName).ifPresent(op -> Tags.CACHE_OP.set(span, op));
     }
 
     @Override
     public Object afterMethod(EnhancedInstance objInst, Method method, Object[] allArguments, Class<?>[] argumentsTypes,
-        Object ret) throws Throwable {
+                              Object ret) throws Throwable {
         ContextManager.stopSpan();
         return ret;
     }
 
     @Override
     public void handleMethodException(EnhancedInstance objInst, Method method, Object[] allArguments,
-        Class<?>[] argumentsTypes, Throwable t) {
+                                      Class<?>[] argumentsTypes, Throwable t) {
         ContextManager.activeSpan().log(t);
     }
 }